isolate signer code into `signer-common` module
