+++ build.gradle
@@ -3,7 +3,7 @@
 buildscript {
 
     ext {
-        kotlin_version = '1.3.30'
+        kotlin_version = '1.3.31'
         kotlin_serialization_version = '0.11.0'
         coroutines_version = "1.2.0"
 
@@ -15,8 +15,8 @@ buildscript {
         moshi_version = "1.8.0"
         okhttp_version = "3.14.1"
 
-        kmnid_version = "0.3"
-        kethereum_version = "0.75.0"
+        kmnid_version = "0.3.2"
+        kethereum_version = "0.75.1"
 
         current_release_version = "0.0.5"
     }
@@ -42,27 +42,27 @@ plugins {
     id "io.errorlab.gradle.vault" version "0.1.0"
 }
 
-apply plugin: "io.gitlab.arturbosch.detekt"
-
-detekt {
-    version = detekt_version
-    input = files(
-            "$projectDir"
-    )
-    //config = "${projectDir}/detekt.yml"
-    filters = ".*test.*,.*/resources/.*,.*/tmp/.*,.*/build/.*"
-    parallel = true
-    reports {
-        xml {
-            enabled = true
-            destination = file("${project.buildDir}/reports/detekt.xml")
-        }
-        html {
-            enabled = true
-            destination = file("${project.buildDir}/reports/detekt.html")
-        }
-    }
-}
+//apply plugin: "io.gitlab.arturbosch.detekt"
+//
+//detekt {
+//    version = detekt_version
+//    input = files(
+//            "$projectDir"
+//    )
+//    //config = "${projectDir}/detekt.yml"
+//    filters = ".*test.*,.*/resources/.*,.*/tmp/.*,.*/build/.*"
+//    parallel = true
+//    reports {
+//        xml {
+//            enabled = true
+//            destination = file("${project.buildDir}/reports/detekt.xml")
+//        }
+//        html {
+//            enabled = true
+//            destination = file("${project.buildDir}/reports/detekt.html")
+//        }
+//    }
+//}
 
 allprojects {
     repositories {
@@ -92,8 +92,14 @@ allprojects {
     }
 }
 
-apply from: "publishing.gradle"
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
+subprojects { subproject ->
+    afterEvaluate {
+        subproject.tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+            kotlinOptions {
+                jvmTarget = "1.8"
+            }
+        }
+    }
 }
+
+apply from: "publishing.gradle"
