+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':core'
+include ':signer-common'
 include ':jsonrpc'
 include ':test-helpers'
