+++ signer-common/build.gradle
@@ -20,6 +20,7 @@ dependencies {
     api "com.github.komputing.KEthereum:functions:$kethereum_version"
     api "com.github.komputing.KEthereum:crypto:$kethereum_version"
     api "com.github.komputing.KEthereum:crypto_impl_spongycastle:$kethereum_version"
+    api project(":core")
 
     testImplementation "junit:junit:$junit_version"
     testImplementation "com.willowtreeapps.assertk:assertk-jvm:$assertk_version"
