+++ signer-common/src/test/java/me/uport/sdk/signer/KPSignerTests.kt
@@ -1,4 +1,4 @@
-package me.uport.sdk.core
+package me.uport.sdk.signer
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
